[FEAT] study 도메인 솔로 세션 관련 코드 작성 (추후 머지부탁드리겠습니다)

+++ legend-momo-city/src/main/java/com/wanted/momocity/study/infrastructure/persistence/GroupRoomJpaEntity.java
@@ -44,7 +44,7 @@ public class GroupRoomJpaEntity extends BaseTimeEntity {
     private GroupRoom.GroupRoomStatus status;
 
     @Column(name = "max_member", nullable = false)
-    private int maxMember;
+    private byte maxMember;
 
     @Column(name = "deleted_at")
     private LocalDateTime deletedAt;
@@ -56,7 +56,7 @@ public class GroupRoomJpaEntity extends BaseTimeEntity {
         entity.hostUserId = domain.getHostUserId();
         entity.inviteCode = domain.getInviteCode();
         entity.status = domain.getStatus();
-        entity.maxMember = domain.getMaxMember();
+        entity.maxMember = (byte)domain.getMaxMember();
         entity.deletedAt = domain.getDeletedAt();
         return entity;
     }
